Fix compiler problems
GCC didn't seem to recognise format specifiers for unsigned long long, %llu.

- Updated MinGW to latest (gcc-14.2.0-mingw-w64ucrt-12.0.0-r2)
- Enabled stricter compiler warnings: -Wextra, -pedantic-errors, -Wconversion, -Wsign-conversion
- Specified compiler version properly to C17 ISO C
- Corrected format specifiers to be compliant with passed datatypes
- Reverted format specifiers from I64 styling
- Changed util and uci types to correct integer datatypes

+++ util.h
@@ -14,7 +14,7 @@ extern int stopped;
 
 int get_time_ms();
 void reset_tc();
-int input_waiting();
+unsigned long input_waiting();
 void read_input();
 void check_up();
